+++ src/main/java/dev/hcr/hcf/HCF.java
@@ -2,11 +2,16 @@ package dev.hcr.hcf;
 
 import dev.hcr.hcf.commands.admin.EconomyCommand;
 import dev.hcr.hcf.commands.players.BalanceCommand;
+import dev.hcr.hcf.databases.MongoImplementation;
+import dev.hcr.hcf.factions.Faction;
 import dev.hcr.hcf.factions.commands.member.DefaultFactionCommand;
 import dev.hcr.hcf.factions.commands.FactionCommandManager;
 import dev.hcr.hcf.factions.structure.regen.FactionRegenTask;
 import dev.hcr.hcf.factions.types.SafeZoneFaction;
-import dev.hcr.hcf.listeners.FactionListener;
+import dev.hcr.hcf.factions.types.WarzoneFaction;
+import dev.hcr.hcf.factions.types.WildernessFaction;
+import dev.hcr.hcf.listeners.factions.FactionClaimingListener;
+import dev.hcr.hcf.listeners.factions.FactionListener;
 import dev.hcr.hcf.listeners.UserListener;
 import dev.hcr.hcf.utils.backend.ConfigFile;
 import dev.hcr.hcf.utils.backend.types.PropertiesConfiguration;
@@ -22,6 +27,7 @@ import java.util.List;
 public final class HCF extends JavaPlugin {
     private static HCF plugin;
     private final NumberFormat format = NumberFormat.getCurrencyInstance();
+    private MongoImplementation database;
     private FactionRegenTask regenTask;
     private List<ConfigFile> files = new ArrayList<>();
 
@@ -30,6 +36,7 @@ public final class HCF extends JavaPlugin {
         plugin = this;
         loadConfigurationFiles();
         registerCommands();
+        implementDatabases();
         loadFactions();
         registerEvents();
     }
@@ -37,15 +44,21 @@ public final class HCF extends JavaPlugin {
     @Override
     public void onDisable() {
         // Plugin shutdown logic
+        saveFactions();
     }
 
     private void loadConfigurationFiles() {
         files.addAll(Arrays.asList(
-                new ConfigFile("config", this)
+                new ConfigFile("config", this),
+                new ConfigFile("database", this)
         ));
         File directory = new File(plugin.getDataFolder(), "factions");
         directory.mkdirs();
         new PropertiesConfiguration(new File(directory, "faction.properties"));
+    }
+
+    public ConfigFile getConfiguration(String file) {
+        return files.stream().filter(configFile -> configFile.getName().equalsIgnoreCase(file)).findAny().orElse(null);
     }
 
     private void registerCommands() {
@@ -55,14 +68,38 @@ public final class HCF extends JavaPlugin {
         getCommand("faction").setExecutor(new DefaultFactionCommand());
     }
 
+    private void implementDatabases() {
+        database = new MongoImplementation(this);
+    }
+
     private void loadFactions() {
+        database.loadFactions();
+        if (Faction.getWilderness() == null) {
+            new WildernessFaction();
+        }
+        if (Faction.getWarzone() == null) {
+            new WarzoneFaction();
+        }
+        if (Faction.getSafeZone() == null) {
+            new SafeZoneFaction();
+        }
         regenTask = new FactionRegenTask();
         regenTask.runTaskTimerAsynchronously(this, 20L, 20L);
-        new SafeZoneFaction("safezone");
+
+    }
+
+    void saveFactions() {
+        for (Faction faction : Faction.getFactions()) {
+            database.appendFactionData(faction.save());
+        }
     }
 
     private void registerEvents() {
-        Arrays.asList(new UserListener(), new FactionListener()).forEach(listener -> Bukkit.getPluginManager().registerEvents(listener, this));
+        Arrays.asList(new UserListener(), new FactionListener(), new FactionClaimingListener()).forEach(listener -> Bukkit.getPluginManager().registerEvents(listener, this));
+    }
+
+    public MongoImplementation getMongoImplementation() {
+        return database;
     }
 
     public FactionRegenTask getRegenTask() {