+++ src/main/java/dev/hcr/hcf/factions/commands/FactionCommandManager.java
@@ -2,6 +2,7 @@ package dev.hcr.hcf.factions.commands;
 
 import dev.hcr.hcf.factions.commands.captain.FactionInviteCommand;
 import dev.hcr.hcf.factions.commands.captain.FactionWithdrawCommand;
+import dev.hcr.hcf.factions.commands.coleader.FactionClaimCommand;
 import dev.hcr.hcf.factions.commands.leader.FactionCreateCommand;
 import dev.hcr.hcf.factions.commands.member.*;
 import dev.hcr.hcf.factions.commands.staff.FactionTestCommand;
@@ -18,6 +19,9 @@ public class FactionCommandManager {
         // CAPTAIN COMMANDS
         new FactionInviteCommand();
         new FactionWithdrawCommand();
+
+        // COLEADER COMMANDS
+        new FactionClaimCommand();
 
         // LEADER COMMANDS
         new FactionCreateCommand();