+++ src/main/java/dev/hcr/hcf/factions/commands/staff/FactionTestCommand.java
@@ -1,11 +1,15 @@
 package dev.hcr.hcf.factions.commands.staff;
 
+import dev.hcr.hcf.factions.claims.Claim;
 import dev.hcr.hcf.factions.commands.FactionCommand;
-import dev.hcr.hcf.utils.backend.ConfigurationType;
+import dev.hcr.hcf.factions.types.PlayerFaction;
+import dev.hcr.hcf.users.User;
+import dev.hcr.hcf.utils.CC;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
-import org.bukkit.configuration.file.FileConfiguration;
+import org.bukkit.entity.Player;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class FactionTestCommand extends FactionCommand {
@@ -16,12 +20,27 @@ public class FactionTestCommand extends FactionCommand {
 
     @Override
     public void execute(CommandSender sender, Command command, String label, String[] args) {
-        ConfigurationType configuration = ConfigurationType.getConfiguration("faction.properties");
-        if (configuration == null) {
-            sender.sendMessage("Could not find file!");
+        if (!(sender instanceof Player)) return;
+        Player player = (Player) sender;
+        User user = User.getUser(player.getUniqueId());
+        if (user.getFaction() == null) {
+            player.sendMessage("No faction!");
             return;
         }
-        sender.sendMessage(configuration.getString("team-damage"));
+        PlayerFaction playerFaction = (PlayerFaction) user.getFaction();
+        List<String> message = new ArrayList<>();
+        message.add("&7&m--------------------------------------");
+        if (playerFaction.hasClaims()) {
+            message.add("&7Claims: &aTRUE");
+            for (Claim claim : playerFaction.getClaims()) {
+                message.add("  &7Position 1: &c" + claim.getCuboid().getPoint1().getX() + "," + claim.getCuboid().getPoint1().getZ());
+                message.add("  &7Position 2: &c" + claim.getCuboid().getPoint2().getX() + "," + claim.getCuboid().getPoint2().getZ());
+            }
+        } else {
+            message.add("&7Claims: &cFALSE");
+        }
+        message.add("&7&m--------------------------------------");
+        message.forEach(msg -> player.sendMessage(CC.translate(msg)));
     }
 
     @Override