+++ src/main/java/dev/hcr/hcf/listeners/factions/FactionListener.java
@@ -1,18 +1,27 @@
-package dev.hcr.hcf.listeners;
+package dev.hcr.hcf.listeners.factions;
 
+import dev.hcr.hcf.factions.Faction;
 import dev.hcr.hcf.factions.structure.Relation;
+import dev.hcr.hcf.factions.structure.SystemFaction;
 import dev.hcr.hcf.factions.types.PlayerFaction;
+import dev.hcr.hcf.factions.types.WarzoneFaction;
 import dev.hcr.hcf.users.User;
 import dev.hcr.hcf.users.faction.ChatChannel;
 import dev.hcr.hcf.utils.CC;
 import dev.hcr.hcf.utils.backend.ConfigurationType;
-import org.bukkit.configuration.file.FileConfiguration;
+import org.bukkit.Bukkit;
+import org.bukkit.Location;
+import org.bukkit.entity.Entity;
 import org.bukkit.entity.Player;
+import org.bukkit.event.Event;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
+import org.bukkit.event.block.BlockBreakEvent;
+import org.bukkit.event.block.BlockPlaceEvent;
 import org.bukkit.event.entity.EntityDamageByEntityEvent;
 import org.bukkit.event.entity.PlayerDeathEvent;
 import org.bukkit.event.player.AsyncPlayerChatEvent;
+import org.bukkit.event.player.PlayerMoveEvent;
 
 public class FactionListener implements Listener {
 
@@ -65,5 +74,64 @@ public class FactionListener implements Listener {
         PlayerFaction faction = (PlayerFaction) user.getFaction();
         // TODO: 1/30/2022 Get the current location of the players death, see if this location is a faction and check to see if it has dtr multipliers
         faction.decreaseDTR(ConfigurationType.getConfiguration("faction.properties").getDouble("dtr-multiplier"));
+    }
+
+    @EventHandler
+    public void onMoveEvent(PlayerMoveEvent event) {
+        Player player = event.getPlayer();
+        Location to = event.getTo();
+        Location from = event.getFrom();
+        if (to.getBlockX() != from.getBlockX() || to.getBlockZ() != from.getBlockZ()) {
+            // TODO: 1/30/2022 create custom events for entering and leaving territory
+            Faction factionTo = Faction.getByLocation(player.getLocation());
+            if (factionTo != null) {
+                player.sendMessage("Entered: " + factionTo.getName());
+            }
+            Faction factionFrom = Faction.getByLocation(player.getLocation());
+            if (factionFrom != null) {
+                player.sendMessage("Left: " + factionFrom.getName());
+            }
+        }
+    }
+
+    @EventHandler
+    public void onBlockBreak(BlockBreakEvent event) {
+        Player player = event.getPlayer();
+        Location location = event.getBlock().getLocation();;
+        event.setCancelled(!canDamageTerritory(player, location));
+    }
+
+    @EventHandler
+    public void onBlockPlace(BlockPlaceEvent event) {
+        Player player = event.getPlayer();
+        Location location = event.getBlock().getLocation();
+        event.setCancelled(!canDamageTerritory(player, location));
+    }
+
+    private boolean canDamageTerritory(Entity entity, Location location) {
+        if (!(entity instanceof Player)) {
+            //  if the "entity" who is damaging the territory is not a player we will prevent it from doing damage
+            return false;
+        }
+        Player player = (Player) entity;
+        Faction factionLocation = Faction.getByLocation(location);
+        if (factionLocation instanceof WarzoneFaction) {
+            Location spawn = new Location(Bukkit.getWorld("world"), 0, location.getY(), 0);
+            double distance = spawn.distance(location);
+            if (distance > 300) {
+                // TODO: 1/31/2022 make build radius configurable
+                return true;
+            }
+        }
+        if (factionLocation instanceof SystemFaction) {
+            return false;
+        }
+        User user = User.getUser(player.getUniqueId());
+        PlayerFaction playerFaction = (PlayerFaction) user.getFaction();
+
+        if (factionLocation.getName().equalsIgnoreCase(playerFaction.getName())) {
+            return true;
+        }
+        return false;
     }
 }